+++ src/app/api/auth.api.model.ts
@@ -1,3 +1,5 @@
+import { AccountFragment } from "./kamu.graphql.interface";
+
 export interface PasswordLoginCredentials {
     login: string;
     password: string;
@@ -5,4 +7,9 @@ export interface PasswordLoginCredentials {
 
 export interface GithubLoginCredentials {
     code: string;
+}
+
+export interface LoginResponseType {
+    accessToken: string;
+    account: AccountFragment;
 }