+++ CHANGELOG.md
@@ -1,16 +1,19 @@
 # Changelog
-
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added a feature for changing account email
+- Added `redirectUrl` parameter in the URL for login page
+
+## [0.38.0] - 2025-01-27
 ### Changed
 - Replaced the package for custom svg icons with the MatIconRegistry
 - Replaced the lodash package with custom helpers
 - Replaced the moment.js package with date-fns
-
 
 ## [0.37.0] - 2025-01-17
 ### Added
@@ -19,7 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Environment variables and secrets: made `key` field unavailable for editing
 - Improved representation of retractions and corrections in data table
-
 
 ## [0.36.0] - 2025-01-08
 ### Added